backend/apps/etl_sqlite/pyproject.toml
[tool.poetry]
name = "sqlite-to-postgres"
version = "0.1.0"
description = ""
authors = ["Nikita <zigman.nikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.7"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
